refactor: 인증 및 보안 관련 클래스 수정
- JwtProvider 및 CustomUserDetailsService 관련 코드를 수정하였습니다.
